Move from overleaf (#14)
* Added new file structure

* Added code files to newly created src directory

* Copy and pasted from Overleaf

* Added words to known.py and played with main.tex

Known words were not on this branch.

Deleted some of main.tex to see if that would solve the "recipe terminated with error". It didn't.

Recieve the following message on saving:
latexmk: The script engine could not be found.
latexmk: Data: scriptEngine="perl.exe"

Have downloaded Strawberry Perl and will try to use that as the script engine.